add extended legend for nodelist acc metrics, move nodelist paging select
