+++ web/templates/config.tmpl
@@ -8,6 +8,7 @@
 {{define "javascript"}}
     <script>
         const isAdmin = {{ .User.HasRole .Roles.admin }};
+        const isSupport = {{ .User.HasRole .Roles.support }};
         const isApi = {{ .User.HasRole .Roles.api }};
         const username = {{ .User.Username }};
         const filterPresets = {{ .FilterPresets }};
